+++ src/components/input/input.ts
@@ -255,7 +255,9 @@ export default class SlInput extends LitElement {
 
   @watch('value')
   handleValueChange() {
-    this.invalid = !this.input.checkValidity();
+    if (this.input) {
+      this.invalid = !this.input.checkValidity();
+    }
   }
 
   render() {